% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separator.R
\name{separator}
\alias{separator}
\title{Separates a clip into stems (vocals, bass, drums, others)}
\usage{
separator(audio_path = NULL, n_stem = 2)
}
\arguments{
\item{audio_path}{The file path where the music are located.}

\item{n_stem}{A number [2, 4 and 5] to indicates a flavour of separation, where:
\itemize{
  \item{2: Vocals (singing voice) / accompaniment separation}
  \item{4: Vocals / drums / bass / other separation}
  \item{5: Vocals / drums / bass / piano / other separation}
}}
}
\value{
Creates a directory with files in .wav format
}
\description{
Separates a clip into stems (vocals, bass, drums, others)
}
\examples{
\dontrun{
library(spleeteR)
separator(audio_path = 'path_to_music/music.mp3', n_stem = 5)
}
}
